Extract specific logic into separate file
